requirements.txt
numpy
gast>=0.3.3
tensorflow>=2.7,<2.12
scipy
scikit-learn
matplotlib>=3.2
imageio
opencv-python-headless
shapely
umap-learn
seaborn<0.12
pandas
pyvips>=2.1.11
fpdf2
cplex
cvxpy>=1.1.15
lifelines
torch>=1.9.1,<1.13
torchvision
pretrainedmodels
gitpython
scikit-image
tqdm
click
protobuf<=3.20.2
tensorboard
crc32c
h5py
tabulate
saliency
rasterio
smac==1.4.0
ConfigSpace
pyarrow
ninja
tensorflow_probability<0.20
rich
pillow>=6.0.0
imgui
pyopengl
glfw
saliency
parameterized
pyperclip
requests
cucim
versioneer
cellpose
zarr
fastai